Removed old examples - added instant register update if write was successful
