+++ DocBuilder/DocBuilder.csproj
@@ -1,10 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
+
+	<IsPublishable>false</IsPublishable>
+	<IsPackable>false</IsPackable>
     <OutputType>Exe</OutputType>
     <TargetFramework>net6.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
+  
   </PropertyGroup>
 
   <ItemGroup>
